Got fragments working correctly.
Displays as each screen separately on phone and in side-by-side panes on
tablet.  Reworked the add movie activity as it was only pulling part of
the layout from the fragment.  Got the back button working to go back to
the previous fragment on phone.  Almost done adding fields in add movie
and need to do some additional styling.